Experimental addition of concurrency.
Adds (a possible) semantics for concurrent execution to layers. Transitivity of
layer refinements is proved, except for initialization. There is also
unnecessary use of functional extensionality that should be removed.

The semantics is somewhat inspired by CSPEC, but differing in a few ways:

- No thread IDs. Thread IDs could be simulated by having parent threads
bump a shared counter and then substituting the value into the child
before spawning. Alternatively, we could modify the semantics of layers
to add a "spawn_step" command that would modify state during a spawn.

- No trace semantics. The observational behavior of a program is whatever the
"parent thread" (i.e. the first thread) evaluates to.  It seems to me that
this fits slightly better with how things were done in Argosy than trace
semantics would. I think CSPEC trace semantics can be simulated using a special
piece of state (which would collect the list of events), if I understand
correctly.

- No atomic wrapper.

There are also some differences from Argosy:

- Recovery runs a *sequence* of programs, and compilation of recovery just
appends the compiled abstract recovery sequence to the end of the concrete
layer's. This means that child threads forked by a concrete layer's recovery
procedure are stopped before turning over to the abstract layer's procedure.

- To prove something is a LayerRefinement, you have to show that the behavior
of *whole programs* is preserved, rather than just per-operation results.  I
don't see how to phrase things in terms of per-operation results without baking
in the concurrency reasoning style one is going to use. So, obviously this
means more of the hard work is pushed to an obligation and the framework does
less.