Add traces to the concurrent semantics.
Traces are incorporated as part of a layer's state. When defining a layer, one
specifies a projection State -> list Event.

Very little is assumed about this function, other than that crash steps
preserve the trace. However, one probably wants to show that the top/bottom
layers of the stack have the property that this list is monotonically growing
as steps are taken.

The correctness criteria for layer refinement now require that
if we start executing from a state where the abstraction relation holds, then
after each step of execution of the concrete program, there exists a partial
execution of the abstract program which has an identical trace (we may want to
change this to be a *prefix* rather than equality). This trace property
continues to hold after crashes throughout recovery.

This property is a little odd unless the traces monotonically grow because the
partial execution of the abstract program that we match with after each step
need not be the same.