Some progress on modeling data structures
Got stuck on dealing with type indices; couldn't write down a strong
update on well-typed state of the form [forall T, IORef T -> T] since
the indices [T & IORef T] would need decidable equality and I'm not sure
about assuming that. In theory it's fine, but we can't actually compare
heterogeneous IORefs in Haskell (the only facility I know of to do this
is Data.Typeable, and IORefs don't implement Typeable). This is a
strange issue since I'm having trouble giving a _model_ of IORefs, not
assuming pure functions to use them.

Solved this problem by modeling IORefs using type codes. Reads are an
Err in the semantics if the "guessed" type isn't correct.