Implement non-atomic operation plan for IORef
The idea: split writes into two parts, begin and finish. Begin
registers that a thread is arbitrarily mutating the object, and during
this time any concurrent Begin's and reads error. Finish does the actual
operation (apparently atomically). This approximates a non-linearizable
operation as simply being unsafe for concurrent access, which is exactly
the contract for many "non-thread-safe" data structures, like all the
mutable data structures we're using in Haskell (IORef, Vector, and
HashTable).