Implement mutable arrays
After much searching I finally found a package, dynamic-mvectors, that
gives mutable arrays with reserved capacity for amortized O(1) appends.
It's a wrapper around MVectors from the vector package that tracks a
dynamic length and capacity as references (the length of an MVector is
an immutable part of the reference).