Prove the remaining encoders
Turns out it was just a matter of some simple rewriting automation.

Note that the proofs aren't really based on verified combinators, so
adding more correctness properties will require fixing the automation
(and this won't be as easy as updating combinator proofs, since the new
correctness conditions are probably structurally different).