Add pure part of slice model
Built on top of IORefs; this unfortunately allows coercing a slice to an
IORef and vice versa, but we have two options:
- Give this the expected semantics of reading the first element. To
actually run the code we'll need some unsafe.* operations in Go, or in C
this is easy. In practice we just won't use this feature so we don't
need to worry about the equivalent Go.
- Allocate from two different spaces so p.(slice.ptr) is an invalid
pointer except through slice operations and same for slices fabricated
from pointers