Add goose mail server
Link is still annoying to have in the filesystem model, but maybe it's
not so bad with some linearity restrictions or after wrapping all the
indirections in helper specs. That said, until it's necessary I'll still
write a model that doesn't have inodes (links will be allowed, but only
after closing all fds, which would mean it's impossible to modify a file
after creating a hard link).