Rough proof of Goose refinement adeq. (w/ 1 admit)
Besides cleaning up the messy proof and supporting globals, the remaining part
that needs to be handled is that during a crash, the LockStatuses that we use
to detect readdir/delete races in the semantics get reset to "Unlocked".

This ensures that recovery can run "readdir" without triggering an error.  We
need to pass the fact that *all directories* are unlocked to the precondition
of recovery, but this means the invariant needs to be strengthened to recognize
the fact that we don't remove/add directories during execution.