Split Pickup into two ops.
At the top layer for the MailServer, Pickup previously did two things: it took
an atomic snapshot of current messages and created a slice of slices containing
all the messages in the inbox. This commit splits Pickup into separate
primitive operations for these two things.

The reason is that we do not have a good way to prove that the implementation
of Pickup atomically does these two steps. The true linearization point for the
implementation is the moment we call readdir on the user's inbox and get a list
of messages. But at that point we haven't actually created the slice of slices
with all messages. (Moreover, when we do finish creating that slice, the inbox
will contain new messages).

In actuality, it is fine that the slice is not yet fully created at that
linearization point, because no concurrent thread can forge a pointer to our
partially initialized slice and observe its contents. However, we don't have a
good way with the current shallow embedding to prove that this forging is
impossible/would trigger undefined behavior.

Note that this issue did not arise in CMAIL because there Pickup returned a
pure list of values rather than a pointer.