Split deliver into two sub-operations.
The first operation read-locks the input slice containing the message, while
the second actually does the delivery and unlocks the slice.

This is equivalent in some sense to the previous version, because any racing
write modifying the input would have triggered undefined behavior, since there
was a check to see that the input slice was read-lockable, but I think it is
easier to reason about this version.