Generalize refinement proofs to allow more flexible forms of translation.
This commit generalizes the generic adequacy theorems so that the notion of
compilation/translation between layers doesn't have to be given by a
deterministic function from procs at one layer to procs at the lower layer.

The Goose-specific version is updated to exploit this to handle translations
where we map a sequence of operations to another sequence, specified by giving
a compilation relation between procs.

It was convenient to also generalize certain intermediary proofs to not
necessarily require that refinement imply *exact* equality of return values, or
even that return values be of the same type. The final derived theorems for
Goose and ExMach do use a notion of refinement that implies exact equality of
return values, but going forward we will need this more flexible form of
refinement to reason about data abstraction/private data types.