Toy around with a shallow Goose type system
Don't actually have the type system, but have the surrounding parts: a
model of types and an interpretion of types into Gallina. Dependent
types go surprisingly far in writing ordinary Coq code and getting it to
type check as [rep ty] for some type code ty; hopefully we can make the
type system syntactic enough that type judgements are similarly trivial
to construct.