Define recovery weakestpre.
Recovery weakest pre takes an expression and a recovery procedure to run in
case the expression crashes. It's defined as a guarded fixed point of repeated
crash weakest pres.

We then have a straight-forward proof of what Argosy calls the "idempotence"
lemma: to prove a recovery weakest pre, it suffices to prove appropriate
crash weakest pres for the expression and the recovery procedure.

This definition is not perfect yet I think, but the fact that we get this
resemblence to the Argosy stuff suggests to me that this is a better way to
phrase things.