Add u32 value constructor
Since u64 and u32 are now proper newtype inductives, we can use
coercions to create them and are back to the nice support heap_lang has.
Integers are by default coerced to u64, which means #1 defaults to a
u64. Getting a u32 is currently awkward; probably we should make U32 be
a function from Z to u32 (and change the name of the constructor).