Clean up the struct field typing theorems a bit
Pushed a little bit of the work into the type system and I think the
result is a bit cleaner while also proving a theorem about field offsets
in isolation, which will make proving storing into fields a simple
extension of storing a general type.

The remaining messiness is no longer about list theorems but instead
dealing with the fact that the struct code carefully builds non-empty
lists with left associativity, which is terrible for reasoning about.