Start proving slice triples
I want to hide away the physical representation of slices (that is, the
value (#l, #sz)) - the current implementation does this, but I think the
way it's done is awkward and I feel like there's a more standard/more
clever approach which is easier to use. Heap lang code seems to avoid
this by exposing the representation, which is just a loc 90% of the
time.

Whatever solution we use should hopefully generalize to Go structs (and
we can generate code for representing them in proofs from goose).