Factor out proof of new is_log after append
Only one admit in append proof, for an overflow case (new blocks could
"fit" on disk via an overflow in 1 + sz + len(bks)).

The code is actually wrong (for impossibly large logs), maybe the
easiest thing is to fix it. While this is annoying, maybe we should just
start being rigorous about doing bounds checks in a way that handles
overflow, for the sake of proofs.