Implement some support for disk interface
A bit of a hack - disk.Disk is modeled as an interface with no
information content, so the model is still global but there's a value
that refers to the opaque disk. This might be a sound model of any usage
where there really is only one disk all callers actually use.