Track slice capacity
Track slice capacity with appropriate non-deterministic extra allocation
(including at `make` time since the translator allows but ignores any
extra capacity argument). I believe this faithfully models Go's intended
semantics, although the specification is rather vague about how
capacities are updated.

Note that this fixes the unsoundness from not having deallocation in
SliceAppend and removes the need to have deallocation at all.

This complicates proofs in a subtle way: we now have a "large" is_slice
that includes points-to facts for the extra capacity and a "small"
is_slice_small that only has the slice's values. It turns out that
is_slice_small is what you want in most cases because it's a weaker
assumption - it doesn't require the caller to precisely track capacity,
which is what most code does.

The annoying thing is that now sometimes you have to weaken is_slice,
which just adds annoying work because the generic theorems need to be as
strong as possible.

We can fix all of this with the right slice library and probably need to
start putting a layer of abstraction so it's easier to use the slice
library in different ways and have the proofs nicely layered (right now
callers often unfold is_slice, which is bad for making deep changes).