Fix test naming scheme
Changing Go function names to indicate failure means that the translated
functions change too. Without modifying goose to handle this
specifically, the solution is just to re-run goose on the semantics
package whenever a test's pass/fail status changes.

This is non-ideal because Coq will let a function fail if the function
doesn't exist. We're prone to false-negatives this way, but not
false-positives, so maybe this is fine for now.