Switch to a simpler struct format
Value representation is now a right-associative, unit-terminated list of
fields. On the heap we do not store the units, so we do not have to
allocate space for the extra "field". It's unclear if this optimization
is really necessary, but it certainly makes things uniform and makes
indices easy to calculate, and stable to re-grouping structs.