Add fraction to is_slice
This is doable but gets a bit out of hand. I'd like to factor this out
generically as read-only vs full ownership, but I'm concerned that
somewhere we'll want to "borrow" read-only and then restore full
ownership (in Rust terminology). To support that pattern we essentially
need fractions or counting permissions.

Fractions have the distinct convenience that full ownership is literally
a specialization of fractional ownership, definitionally, so read-only
theorems can seamlessly be used with full ownership.