Versions of crash inv with easier step counting parameters.
The wpc "k" parameter, representing the number of laters you can do per step in
the wpc, in the premises and conclusions of the crash_inv results was
complicated and involved some tricky doubling/adding on of extra steps. This
commit defines a function LVL : nat -> nat, and then states the crash_inv
lemmas in terms of this LVL, which handles the doubling/adding under the hood.
Each application of a crash_inv lemma then just decrements the LVL count by 1
or 2.