Implement general read-only mapsto
From a user perspective, this creates a magical points-to transformer
written [readonly (l |-> v)] that is persistent and supports extracting
a points-to fact with a fraction.  The support is generic to handle
typed points-to (struct_mapsto, l |->[t] v) as well as the field
points-to (struct_field_mapsto, l |->[S :: f] v). We use typeclass
machinery to handle these without creating additional syntax for each
read-only variant.

wp_loadField knows about [readonly (l |->[Struct.S :: f] v)] and will
load from this fact if it's in the persistent context.

This implementation was inspired by AsFractional.