Perform some major surgery on wal invariant
Hopefully a lot more of it is in the lock invariant now, to minimize use
of the global invariant.

The groupTxns plan is now more or less in place, but it seems almost
arbitrary what constitutes a groupTxn. Maybe its simple because all
we're doing is tracking what subset of the abstract state corresponded
to nextDiskEnd?