Strengthen wp_Write_fupd with a later
This is needed to allow the caller to open an invariant to get the disk
points-to and then close it a step later.

This lets us drop the requirement that the P over the circ state is
Timeless (which was probably true, but maybe we shouldn't keep making
that assumption).