Precisely track capacity for memLog
Needed to append to it, and unfortunately our support for slice capacity
is not great. It would have been nice to have better predicates for
talking about subslices precisely without involving the model of slices
or u64s everywhere, but capacity is just hard to correctly fit into the
reasoning regardless.