Rewrite allocator spec in terms of used set
The code now computes a free set from a domain (which doesn't change)
and a used set, which we reflect in the spec. The key is that recovery
now imposes an obligation on the used set being the same as when we
crashed, rather than the free set, which should be easier to prove in
terms of tracking all the allocated blocks.

In addition, we develop a library range_set for constructing a set for a
range of integers, which was previously not precisely specified in the
wp_freeRange proof. This is now more important since it's internal to
the library.