Stop trying to strip laters in wp_pure so much
Do a coarse check for a hypothesis that starts with a later (which will
actually get stripped), and if one is not found instead apply a tac_
theorem that does not change the environment and does not resolve
MaybeIntoLaterNEnvs.

We were spending considerable time resolving this typeclass, and this
almost entirely eliminates that.