WIP defining a "subresource laterable" modality.
Work in progress to define a modality that captures that something is provable
from a subset of resources that are laterable. Goal is to avoid having to
subscript wpc by k steps using this modality for the crash condition.

Also proves a lemma to derive an accessor from an entailment, if the fact
you're accessing is exactly ownership of a resource.