Generalize bi_except_0 and Timeless.
Iris uses the "bi_except_0" modality in the definition of fupd, which allows
stripping a later from Timeless things during a wp proof.

This commit defines `bi_atleast k` and `AbsolutelyTimeless`. The idea is that
one can strip k laters from an AbsolutelyTimeless assertion when proving a goal
with an `bi_atleast k` modality. (`bi_atleast 1` is the same as `bi_except_0`.)
Essentially everything that was `Timeless` in Iris is actually
`AbsolutelyTimeless` (even though, strictly speaking the former does not imply
the latter).

A few lemmas and instances for `bi_except_0` and `Timeless` have analogs for
these new definitions that are still unproven.