Prove byte decoding
This is actually somewhat interesting, because the decoder seemingly
gives away private state by exposing an internal pointer to its own byte
slice. The way the decoder state is set up it remembers the entire slice
(not just the remainder after the offset), but with an arbitrary
fraction q, so we just cut that fraction in half and give the caller the
other half.