Introduce a HoldsAt notion
Expresses that a predicate holds with a fixed domain.

This is a much nicer version of the old liftable2.v which is internal to
the logic, which is considerably stronger because we can show that if P
holds then there exists a domain where it holds, where the exists is
within the logic and hence assumes P holds of some resources.