Establish that P holds over exactly mT
We can unify the existential map in HoldsAt P (ptsto γ) with the map_ctx
map, proving that P holds over exactly the transaction map. This is
necessary to restore the txn_system_inv, which owns the async state.

The commit has a bunch of other changes fixing the auth_map
infrastructure to export a ptsto that generalizes read-only and mutable
ownership (with an option frac). I also introduced a notion for just the
wand part of HoldsAt, which fixes the map used, since that's what we
need after unifying the maps.