Start using MakePreparedRequest;
need to actually figure out what post we want on increment_core and
MakePreparedRequest. We should only use up γPost at the commit point
(when we make the reply table durable). Maybe make post of
increment_core be a fupd that takes the Post out of the request
invariant; the fupd probably will require a witness that the seqno has
increased past the req's seqno, so that we can deal with crash
condition.