Proof of wp_EncMap, with a more sensible spec;
don't want to bother trying to make marshalling a map deterministic in
the presence of a (should be) non-deterministic MapIter, so
defined a `has_map_encoding (m:gmap u64 u64) (r:Rec)` that says that `r`
encodes the map; can use by writing e.g.
\exists rmap, (has_map_encoding m rmap) /\
        has_encoding data_on_disk ([EncUInt64 foo] ++ rmap ++ [EncUInt64 bar])
to keep track of the fact that a map was serialized to durable storage.

Will want to put in marshall (instead of the durable kvserver) at some point