Finished wpc_MakeFreshIncrClerk;
only admit in incr_proxy is in merging the if proof branch, and
implementing and proving a recovery function.

The recovery will need access to the file ownership in _own_ghost, so
might be able to merge _own_ghost and _own_durable into one big
_own_durable, where the ghost state and durable ownership stay together.

On the other hand, the point of splitting the _own_durable and
_own_ghost is because idempotent core can give fupd for updating
_own_ghost, which is separate from updating _own_durable by calling
`WriteDurableFooServer`, so might not be able to merge.