Proved quiesce_idemp, modulo timelessness of resources;
Given an invariant that has a quiesce_fupd, we want to get another
quiesce_fupd that grants the resources that the in the initial
quiesce_fupd.

There's two ways of doing it. The first is by using a fmcounter
lower-bound fact when you want to turn a double quiesce_fupd into a
single quiesce_fupd.

The other option is to "conditionally delay" specialization of
quiescable_pre to a particular sequence number until the commit point by
putting (quiesceable_pre ∧ quiesce_fupd) in the RPCRequest_invariant.

i.e. put in ( (∀ seq, fm[[cseq]]≥ seq -∗ quiesce_fupd seq) ∧ quiesce_fupd specific_seqno )
That's equivalent to putting in the forall statement and the fm[[cseq]]≥
fact for the specific_seqno, and letting the server specialize at
commit-time, which might be easier.