Progress on "recursively" quiescing rpc_atomic_fupd;
Dealt with later issue by using make_laterable, and keeping the timeless
resources Q in the invariant, and keep a wand to get the fupd out of the
timeless Q.

Need to write a version of HandleRequest_is_rpcHandler that takes the
fupd-granting-wand as well as the request invariant.

Gave up on trying to make a semantic invariant approach work.

Lots of junk to remove.