Figured out 2pc invariant (hopefully);
The initial plan was to have four states, unprepared, prepared,
committed, and aborted.
Then, participants transition from unprepared to prepared, while the
transaction coordinator would transition from (all) prepared ->
committed, as well as unprepared -> abort and prepared -> abort.

That approach was troublesome, and I couldn't tell why at first.
The problem is that if the TC transitions from unprepared to aborted,
there might still be a in-flightprepare request that never reached its
destination, and so a participant might run the prepare method and want
to logically go from Unprepared -> Prepared, but would be unable to
because the TC already aborted. Indeed, even after the TC aborts, a
participant might actually prepare if stuff gets reordered.

So, now, each participant makes the logical transition from
unprepared -> aborted or prepared -> aborted, instead of the TC doing
it, and it seems like it might work. Each participant keeps a "finished"
token that it keeps for all TIDs that it has not aborted or committed
(including TIDs that haven't been seen at all yet). In order to
transition to the Aborted state, you must have this finished token, so
the TC is now prevented from making the abort transition for a
participant.