Stuck in simplified 2pc plan;
intended spec for Participant__PrepareIncrease would give
  is_prepared (k |-> v) (k |-> v+amnt)
as postcondition, by allocating the invariant itself. But now the
coordinator is not committing to whether it wants to Increase or
Decrease for this tid. So, it's possible for coordinator to call
PrepareDecrease concurrently with PrepareIncrease, and the naive spec
would give is_prepared (...) (k |-> v+amnt) and is_prepared (...) (k |-> v-amnt),
which is bogus. Previously, this wasn't an issue because the coordinator
was allocating the 2pc invariant, and would essentially commit to one of
the two operations for that tid and pid.

One option is to once again have the coordinator alloc the invariant,
and once again have an explicit unprepared state (which is pretty much
what it was before).
  The annoying part of that used to be that the coordinator
didn't know exactly what resources to expect, so there was that annoying
quantification over the oldvalue of that key, as well as explicit ghost
state mapping (tid,pid) -> oldvalue.
  However, I think I see how to get rid of that explicit ghost state by
keeping a copy of the invariant on the participant side, which I
previously wasn't doing.