Mostly done with aof__Append(), except an actual bug that came up.
Currently, we use durableLength to keep track of the number of bytes that
have actually made it to disk. We need to know at the end of
Append(newdata)
the number of bytes that must be made durable for the caller to know
that newdata was put on disk, but I currently compute that simply as
"durableLength + len(membuf)". When the AtomicAppend()ing thread resets
membuf to empty, it doesn't increase durableLength (that wouldn't be
safe, since the stuff hasn't actually been made durable). So, we
actually do need to have a second counter "length" keep track of the
entire "logical", but not necessarily durable log! I had incorrectly
thought we could get rid of it and just do `durableLength +
len(membuf)`.