Finish shard Start() proof;
Need to do a bit of copy/paste in move_shard_proof, after which the
biggest missing thing is marshalling for map[uint64][]byte.

The coord server is "trivial" in the sense that it only invokes
MoveShardRPC() on shard servers, which only requires knowing that the
source and the destination servers are both is_shard_server.