Make shard definitions not depend on heapG
Importantly, the fupd passed as a precondition to an RPC should probably
not be an ncfupd -- that would require disambiguating whether we mean
non-crashing (the nc in ncfupd) on the server or on the client.

Since the current library is for non crash executions, we table this
issue by using normal fupds. However, the atomic triple notation uses
ncfupds (because that makes sense throughout the rest of Perennial), so
we introduce an alternate version that uses normal fupds.