Proof of allocating a new proposal,
given the value of some old proposal and given knowledge that all
proposal numbers larger than that old proposal and smaller than the
brand new proposal can never be committed.

The knowledge that the intermediate ones can't be committed will come in
the form of a majority of replicas sending "reject" witnesses for those
proposal numbers.