Add a 'prophetic' async disk semantics
Under this semantics, a disk address contains a pair of blocks:
- the current in memory value
- the on disk/durable value.

On a crash, the disk value stays the same, and the memory value is
replaced by the disk value.  When a write occurs, the in memory value is
updated, and the durable value is non-deterministically updated to the
new value or not.

The key is that the disk value may ONLY change on a write.
Thus, if the disk value is NOT updated by the write, then we have a kind
of "prophecy" that the write never hits disk before crashing.

The tricky thing here is barriers, which normally are supposed to flush
the memory value to disk. But then this could contradict the earlier
prophecy made during the write. We handle this similarly to the way that
Iris handles prophecy variables: by diverging if a barrier is executed
and the prediction of the prophecy was wrong for any address.

It remain to be proved that an adequacy result for this language implies
one for the "standard" previous async disk semantics.