Redo part frontend.Server.FetchAndIncrement with ctr.Clerk.Get().
Right now, the proof requires repetition of proof because of the "if"
statement in the EnterNewEpoch spec. It should be possible to "join"
that if statement somehow to avoid that.  Something like wp_wand, but
for the Specs, not for WPs.

Probably will boil down to functoriality of the Spec Monad endofunctor,
or naturality of the unit transformation.
E.g. wp_wand roughly says that (F : Φ ↦ (WP e { Φ })) is functorial in Φ.