Checkpoint a simplepb proof plan.
In this proof plan, the pb library doesn't know about durability.
The downside of this is that all the protocol ghost state must be
exposed to the user of the pb library so that they can put it into a
crash borrow (someone has to assemble such a crash borrow!). Since they
own it, the user of the pb library must also update the protocol ghost
state. To make this easier, I tried making the callbacks hocap-style, so
the precondition of each callback is a fupd that updates the protocol
state in exactly the right way. Even so, this is a complicated spec.

An alternative plan worth considering is to put the crash borrow in the
pb invariant. It will simplify some things, such as using resources
before the commit point, to e.g. establish inequalities from lower
bound resources. With the current plan, EVERYTHING related to
ghost-state has to be done at the commit point. It will be less
flexible, but I don't see a benefit of the added flexibility of the
current plan.