Coalesce some fupds to make is_proposal_valid simpler; will need to
use later credits or something to deal with the fact that we might open
an arbitrarily large number invariants, and fire fupd from each of them
when committing something. Each fupd will have a later because of the
invariant, and so we could have a ~(num entries in the log) laters.

Prove ghost_commit; reprove ghost_commit except for this later trouble.