Remove bad usage of impredicativity
Each ghost entry in the log previously had an op and a predicate over lists of
ops that was supposed to hold for all prior entries of the log. But this was
done in a way that would have caused universe problems.

This commit instead uses saved predicates that introduce a layer of
indirection. The ghost log entries are now a pair of an op and a gname. The
gname is for a saved_pred_own that stores the predicate that previously would
have been in the entry itself.