Deal with inequivalent physical and ghost map,
due to points-tos being allocated before physical points-tos, by
introducing a "gauge equivalance" and some Proper instances for it. One
of the instances has a lambda term in it, which depends on higher-order
unification and doesn't work well. If I end up cleaning up `map_get`,
I can replace that lambda with `map_lookup_default` or whatever the new
thing will be called.

This is probably overkill for the problem, but it's principled and made
the proof more obvious than it otherwise would have been.