A weird universe constraint is being added for `bad2` even though it's
definitionally equivalent to `bad1`. Can't actually tell what the
primitive constraint is, other than that it collapses the prod universe
and `Logic` universe from iris.bi.

The top-level symptom is pretty unfortunate: if we only have one of
memkv or simplepb in ShouldBuild, everything is ok. If we have both in
either order, there's a universe inconsistency. It seems to indicate
that library is being used in different ways by the two proofs in a
"universe inconsistent" way (though if those libraries were universe
polymorphic, I think there would be no problem).

An immediate fix is to not use universe polymorphism anywhere, and also
not use the normal prod type in `urpc_spec` but make a new `prod'` type
to keep the universes separate.