Goose new mpaxos RSM API and verify it; instead of an `Apply` function,
there is a `TryAcquire()` function which starts an RSM update,
and `TryRelease()` which the user calls to finish and attempt to commit the update.
An analogy:
(`TryAcquire`/`TryRelease` API) is to (mpaxos's old `Apply` API) what
(mvcc's `begin`/`commit` API) is to (mvcc's `Run` API).

This API seemed more convenient to program against for the config
service.