Give up on map library; realized it will be far better to add primitive
maps to GooseLang. The biggest reason for this is that when implementing
maps in GooseLang (using the list library), a `gmap val val` does not
have a canonical GooseLang representation (because the list can be
reordered). As a primitive, the canonical representation could be `Val
LitV (m:gmap val val)`.  That may break the induction principle for
`val`, though.

If maps work well, perhaps we want lists as a primitive too.