Make Var a coercion
It was originally not due to failing the "uniform inheritance criteria",
but this criteria is no longer required.

This should ensure that variables are actually printed as strings,
hiding the Var coercion, which will make code much more readable without
the ugly hack of the `Var'` coercions.

The massive fallout is due to removing many coercions Var' that were
defined for specific type parameters, all of which can now use the
single coercion on Var.