Implement interface comparison more carefully; prove that it's always
safe to compare with nil interface value.

Interface comparisons can panic at runtime if the dynamic types of the
two objects are the same but the type is not comparable. If the types
are not the same, it's OK for the types to be non-comparable. The
built-in EqOp does not support this for the interface value, so
implemented `interface.eq` for it.