Add spec for single sender/receiver channels.
This is a spec for unbuffered channels that supports single sender and single receiver use cases where the number of values sent is known such as join handles, order preserving data pipelines and notifications. I have a proof for this specficiation but I write the proof when the channel model only supported uint64 channels. This spec can be trusted since adding generics won't change the essence of the proof but I am looking to merge this with Admitteds so that in can be used while I make the mechanical changes to the proof to support generics. I am also working on a more powerful spec that this could be built on top of but will likely be more difficult to use, so this might be a long term option.