Add channel model and update prelude/update-goose. (#185)
The prelude needs to contain the channel model but
since the channel model was written using Goose
compatible ghost code, it needs to contain the
prelude. To avoid a circular dependency, I created
a new "translated_prelude" that will contain the
channel model and any future features modeled
this way and moved the existing prelude to
"base_prelude".