Better support for partial method application;
this shows up e.g. in `defer` and `go` statements, and previously
required awkward theorems stating things like "calling method Foo with
one argument `rcvr` yields a `(f : val)` such that `f #arg1 ...`
satisfies the spec of method Foo with receiver `rcvr`".
Now, one can states a normal hoare triple, and if a method is partially
applied, that normal spec will later be applicable.

The trick is that this new `method_call` is essentially an eta-expanded
version of the old one, which means the body won't be unfolded until
more arguments are given.