Pure state machine reasoning for RWMutex is *very* subtle:
e.g. the (wrong) invariant assumes that when a Lock() call starts to
block readers by negating the `readerCount` that no new readers would
begin. But in fact, the `readerSem` might be still positive from the
last time the Unlock() was called, and so the number of readers might
actually go up.

So, need to count the number of active readers very carefully in that
situation by using:
`readerCount = (active readers) + readerSem + (readers blocked because of ongoing Lock)`.

The old goose_lang/src/lib/rwlock `RUnlock` spec is technically wrong
(though one would need to use the spec with e.g. a persistent lock
invariant to exploit this) w.r.t the Go RWMutex. This problematic spec
seems to have been provable because of a liveness bug in the GooseLang
model.