Add a DFractional typeclass (#219)
This is just a start. I didn't add instances for every use of Fractional, just the most basic ones. This does automatically give `UpdateIntoPersistently` for these definitions, which is nice and one of the original motivations here.

We could also try to fully replace Fractional with DFractional, but this would require a little more experimentation and integration with the IPM - deriving Fractional from DFractional automatically seems tricky due to higher-order unification issues.