Implement dynamic, GooseLang-based types and memory operations (#224)
Currently, `load_ty t #l` (and similarly store_ty) take a Gallina type `t` and use it to direct loading a struct from memory. In preparation for generics, we would like `"T"` to be a GooseLang argument so it can be passed to all of these model functions after dynamically looking up the function by name.

This PR first makes `go_type` accessible in GooseLang as a value (the defn and theory dynamic_typing.v), then re-writes `load_ty` and `store_ty` to use a type passed in GooseLang rather than Gallina (the defn and theory dynamic_mem.v). The translation needs a small tweak to pass `#t` for a Gallina type for all current use cases, which will eventually be `"T"` for a generic type argument `T`. (See https://github.com/goose-lang/goose/pull/72 which implements this translation change.)

In the process I did make `arrayT` take a w64 instead of a nat, so that the encoding of `go_type` into `val` is injective.

The file `mem.v` is now the dynamic type notion of memory. The old library is preserved in `static_mem.v` as a nice demo, but no longer used by the translation or any libraries.

This commit also updates the slice, struct, and map libraries to operate on the new dynamic types, so that they can all work with generic types.