Fail in coqdep if a missing file is required
This only affects files that are required from somewhere not built by
src/ShouldBuild.v (since otherwise the build would naturally fail once
that file was needed).

If this triggers, it won't allow building anything else since .coqdep.d
isn't built. It's possible to disable it for a one-off build with `make
COQDEP_ARGS=`.

If this ever becomes a problem we can revert it.