Use `Proof. Admitted.` instead of `Admitted.` to avoid depending on
irrelevant section variable. The old definition of `own_Hasher`
depended on `cryptoffi.GlobalAddrs` (which is empty, and really nothing
should use it) because it was declared with `Admitted.` By first
starting a `Proof.`, the definition ends up only using section variables
which appear in the statement of the "theorem" (i.e. the type).