Treat slice length as a signed integer (#345)
The length of a slice is really natively a signed integer, in the sense
that `len(s)` produces a signed integer and `make` accepts a signed
integer. Our model is sound with unsigned integer because the two types
are represented the same way, lengths are always positive, and in
practice the difference between a 2^63 and 2^64 bound is irrelevant. As
an aside, the real bound on a slice as enforced by the runtime is
typically 2^48 bytes - note that a zero-sized type does not count, so
`make([]struct{}, 1<<63-1)` works perfectly fine.

`word` already mostly handles the automation required here fine. It
cannot prove that `sint.Z x = uint.Z x` (under appropriate assumptions).
This PR adds the relevant theorems but doesn't add a normalization step
for this. It's unclear to me right now how necessary this is, but it
would be certainly a welcome improvement to avoid mysterious `word`
failures that are avoidable.

Relies on https://github.com/goose-lang/goose/pull/131 to add a relevant
unit test.