Integrate grackle into gokv rocq proofs (#360)
Corresponds to [`gokv` PR
#11](https://github.com/mit-pdos/gokv/pull/11), which integrates Grackle
generated go code for nearly all the marshaling in `gokv`.

Work for this PR mostly included re-goosing for the changes in `gokv`,
which also introduced a lot of new `go` packages that need to be covered
by `new/etc/update-goose-new.py`, then updating the `grackle` scripts.

A few bugs were identified and patched in `grackle` itself, including
one fixed name in the templates and how `grackle` detects the "simple"
message optimization.

I hope that I correctly tweaked the CI workflows, but I expect both
`check-goose` and `check-grackle` to fail until the `gokv` PR is merged.