Translate and verify sync.Cond (#361)
There isn't really a proof here since all the interesting work is in the
runtime's `notifyList`, which is just axiomatized. However, translating
it does confirm that sync.Cond is simple.

We could conceivably extend the verification to `notifyList`, although
without a proof of liveness it's not clear what the value would be.