Generate int variants for atomic package from a template (#365)
Workflow is now to write the Uint64 version as normal Rocq code, then
use a simple Python script to do some replacements on that template for
the other integer types (Int64, Uint32, Int32).

The template and auto-generated output are delimited by comment markers
(e.g, `(* BEGIN TEMPLATE *)`).

To update the output in-place, run:
```bash
./etc/int_template.py new/proof/sync/atomic.v
./etc/int_template.py new/trusted_code/sync/atomic.v
```

I only used this for the atomic trusted code and proofs, but it could be
used for other parts of goose.