model with more explicit assumption of collision-freedom
keep track of all hashes, including ones that possibly collide.
separately state assumption about absence of collisions, which is needed
for some of the hash-reasoning lemmas.  almost everything can be Prop as
a result.  inversion/prophecy is a completely separate issue about using
a prophecy model to predict future hashes, unrelated to the question of
whether they are collision-free or not.