Add more automation to wp_if_destruct (#379)
This PR makes a few changes to `wp_if_destruct`:

- The cleanup post-processing now handles all cases of `decide (#b1 =/≠
#b2)` (that is, decide with equality and inequality of booleans), since
they do all come up in various situations. `#true ≠ #false` and the
reverse are now also hints so `auto` can prove them (it can be
challenging to think of how to prove this, especially for beginners).
- Run `try wp_auto` in both branches.
- In `if: (#b)` if `b` is both a boolean and a variable, `destruct b`.
If `b` is a boolean expression it may still make sense to destruct it
with an equation, but we conservatively don't do that and let the user
handle it.

The reason for so many affected proofs is just needing to remove
`wp_auto`, which fails if it doesn't make progress.