Add support for primitive.Mutex (#400)
This PR moves the spinlock implementation and proof to a common
location, then uses it for both `primitive.Mutex` and `sync.Mutex`. The
idea is that `primitive.Mutex` is available to the channel model to
bootstrap goose, without requiring the entire sync library to be
translated with a subset of the Goose primitives. If and when we verify
`sync.Mutex`, `primitive.Mutex` will remain a trusted model of
synchronization.