Use opam to install dependencies (#502)
This PR changes the build to use opam for dependencies.

As a user, this changes your workflow. We no longer use any submodules.
When dependencies change, you'll use `opam install --deps-only .` to
compile and install the new ones into your opam switch (the same one
you're using for Rocq). After that, the build will still use `make`.

The opam file also describes the set of files required to write proofs
using new goose, which are used for the build and install phases when
doing `opam install .` instead of `make`. I wrote a demo of using this
to set up a program proof that depends on perennial via opam (including
goose and caching in CI) here:
https://github.com/tchajed/perennial-example-proof.

---------

Co-authored-by: Upamanyu Sharma <upamanyu@mit.edu>