Add a wp_end tactic (#413)
This is a simple shorthand that makes a big difference in simple,
introductory examples. It also spares typing HΦ.

The tactic is fairly conservative: it does `iApply "HΦ"` and then
further automation only to solve the goal. We could add `iFrame "#*"`
(which helps in a few cases) and we could consider running `iFrame` even
if it doesn't the solve the goal, although technically I think iFrame is
unsafe, especially in instantiating evars.